Solving the problem of working with clients
This is one of my Telegram projects, where I solve a problem for a client-facing organization. Using this code, I remove unnecessary work from workers and simplify the process of ordering products from your business. This project demonstrated my ability to work with a database, with numerous buttons and complex logic. This is clearly a complex task that not everyone can handle.